# src/app/types/employee.ts
// src/types/employee.ts
export type Employee = {
    id: number;
    firstName: string;
    lastName: string;
    position: string;
    phone: string;
    email: string;
    createdAt: Date;
    updatedAt: Date;
  };
  