# src/swagger.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Employee API",
      "version": "1.0.0"
    },
    "paths": {
      "/api/employees": {
        "get": {
          "summary": "Get all employees",
          "responses": {
            "200": {
              "description": "A list of employees",
              "content": {
                "application/json": {
                  "schema": {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/Employee"
                    }
                  }
                }
              }
            }
          }
        },
        "post": {
          "summary": "Create a new employee",
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Employee"
                }
              }
            }
          },
          "responses": {
            "201": {
              "description": "Employee created"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
        "Employee": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer"
            },
            "firstName": {
              "type": "string"
            },
            "lastName": {
              "type": "string"
            },
            "position": {
              "type": "string"
            },
            "phone": {
              "type": "string"
            },
            "email": {
              "type": "string"
            }
          }
        }
      }
    }
  }